Dramatic Speedups
It wasn't clear for me how the minimal change algorithm can speed up the
code. To get the value and weight, I was looping through each bit of the
`BitString` and summing the weights and values.

I didn't take advantage of the fact that the minimal change algorithm
**guarantees** that only one bit can change at each iteration of the
loop. So, the time complexity of my naive solution is O(n2^n), now I managed
to make it O(2^n). So now, n=30 is just 3 seconds on `--release`. I
forgot to mention that the 99% of the time was spent on the progress
bar, which I fixed by updating the progress bar at fixed increments.

This change also prints the items each with their corresponding
weights and values. After the main loop finishes, this also prints the
optimal subset.